+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2024-01-01
+## [v15.12.0] - 2024-01-07
 ### Added
-- In `TimeDisplay` added also the year
+- In `TimeDisplay`, added the display of the year.
+- Added a copy function to `VillaAnnaBVewaesserungPage` and `VillaAnnaMarkisePage` for copying tasks.
+
+### Fixed
+- Fixed incorrect display of task labels in `VillaAnnaMarkisePage`.
 
 ## [v15.11.0] - 2023-12-31
 ### Added